chore: pin flow-bin
Flow-typed reports:

```
UNCAUGHT ERROR: Error: Invalid npm libdef version! It appears to be a non-continugous range
```

Due to an unpinned flow lib def which can be caused by a ranged flow-bin
dependency. There is a
[fix](https://github.com/flow-typed/flow-typed/pull/4018) for flow-typed
in the next version but we run into a few other hurdles so this is just
a quick patch for passing builds.